history.rs: Move some tests to integration
